Do not update status if an exception is thrown.
Some rollback mechanism will be pretty nice here.
If not then make sure all validation are happening before that the status is updated.
It never should happen that the status is half updated and then an exception is thrown.

(this was happening for declare attacker... if a card could not attack then it blew up but the previous cards were updated as attackers. So the front-end will be messed up).