#32: The bug was with the cost not with the summoning sickenss.
The js code was not sending the ability type but always null. This caused that instead of getting the cost of the ability was getting the cost of the card. And for Dread Shade the cost of the card was 3 BLACK while the cost of the ability was 1 BLACK.

The functionality was tested and working on `Jousting Dummy`.