Revert "Swapped to functional components."
I don't think functional component are worth it.
`useEffect` is a mess.

```
  const handleKeyDown = (event) => {
    if (event.key === ' ') {
      if (isContinueEnabled()) {
        props.continueClick()
      }

    } else if (event.key === 'H') {
      props.openHelpPage(true)

    } else if (event.key === 'L') {
      alert('Game Log coming soon!')
    }
  }

  useEffect(() => {
    // FIXME this gets called millions of time
    console.log('registered again')
    document.addEventListener('keydown', handleKeyDown)
    return () => {
      document.removeEventListener('keydown', handleKeyDown)
    }
  }, [props])
```

handleKeyDown needs to use props.
The only way to make handleKeyDown be aware of props is to pass [props] as second argument of `useEffect`. but then whenever props change registration happen again.

This makes no sense at all.
I link the `componentDidMount` it's so clear and the code is just the same.

useState in Login admin is nice... but not worth it!